VEOSS-285: Fix issue with VAT rates that have no decimal part
Due to a combination of the way the Twirl govuk components and json
reads / writes work, there was an issue where checkboxes that
represent a VTA rate that is a whole number (e.g. 20%) would
not be checked when a user revisited that page.  To resolve this, the
rate is written as a string and parsed back into a BigDecimal in
the reads.